Added DashboardContext
