+++ app/javascript/src/common/dummy.js
@@ -1,0 +1,91 @@
+export const ROLES = ["Owner", "User"];
+
+export const DUMMY_CONTACTS = [
+  {
+    firstName: "Oliver",
+    lastName: "Smith",
+    email: "oliver@example.com",
+    role: 0,
+  },
+  {
+    firstName: "James",
+    lastName: "Bond",
+    email: "bond@example.com",
+    role: 0,
+  },
+  {
+    firstName: "Oliver",
+    lastName: "Smith",
+    email: "oliver@example.com",
+    role: 0,
+  },
+  {
+    firstName: "James",
+    lastName: "Bond",
+    email: "bond@example.com",
+    role: 0,
+  },
+  {
+    firstName: "Oliver",
+    lastName: "Smith",
+    email: "oliver@example.com",
+    role: 0,
+  },
+  {
+    firstName: "James",
+    lastName: "Bond",
+    email: "bond@example.com",
+    role: 0,
+  },
+  {
+    firstName: "Oliver",
+    lastName: "Smith",
+    email: "oliver@example.com",
+    role: 0,
+  },
+  {
+    firstName: "James",
+    lastName: "Bond",
+    email: "bond@example.com",
+    role: 0,
+  },
+];
+
+export const TAGS = [
+  "Getting Started",
+  "Onboarding",
+  "User Flow",
+  "UX",
+  "Bugs",
+  "V2",
+];
+
+export const DUMMY_NOTES = [
+  {
+    id: 0,
+    title: "How to claim the warranty?",
+    description:
+      "This is something I found in the Figma design file. So please don't mind",
+    assignedContact: 0,
+    tag: 0,
+    createdOn: "September 24,2022",
+  },
+  {
+    id: 1,
+    title: "Write a blog",
+    description:
+      "This is something I found in the Figma design file. So please don't mind",
+    assignedContact: 1,
+    tag: 0,
+    createdOn: "September 24,2022",
+  },
+  {
+    id: 2,
+    title: "Enjoy your time",
+    description:
+      "This is something I found in the Figma design file. So please don't mind",
+    assignedContact: 2,
+    tag: 0,
+    createdOn: "September 24,2022",
+  },
+];
